结算和打折
