+++ app/src/main/res/layout/bill.xml
@@ -63,7 +63,7 @@
                     android:id="@+id/quantity1"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    android:text="×0"
+                    android:text="0"
                     android:textColor="@color/color_red"
                     android:textSize="20sp"
                     android:textStyle="bold" />
