+++ app/src/main/res/layout/billmain.xml
@@ -14,7 +14,7 @@
         >
 
         <android.support.v7.widget.RecyclerView
-            android:id="@+id/bill_recycleView"
+            android:id="@+id/bill_recycleView1"
             android:layout_width="398dp"
             android:layout_height="618dp"
             android:layout_below="@+id/goodsInfo"
@@ -85,13 +85,9 @@
 
 
     </RelativeLayout>
-
-    <android.support.v4.view.ViewPager
-        android:id="@+id/vp"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent">
-
-    </android.support.v4.view.ViewPager>
+    <include
+        layout="@layout/shop_sub"
+        ></include>
 
 </RelativeLayout>
